Changed money value in Visitor to double and added and tested isAllowedTo method in Playground class

+++ src/main/java/Park/Visitor.java
@@ -3,9 +3,9 @@ package Park;
 public class Visitor {
     private int age;
     private int height;
-    private  int money;
+    private  double money;
 
-    public Visitor(int age, int height, int money){
+    public Visitor(int age, int height, double money){
         this.age = age;
         this.height = height;
         this.money = money;
@@ -19,7 +19,7 @@ public class Visitor {
         return height;
     }
 
-    public int getMoney() {
+    public double getMoney() {
         return money;
     }
